Add assignment code

+++ ct414/ExamServer.java
@@ -9,19 +9,19 @@ import java.util.List;
 public interface ExamServer extends Remote {
 
 	// Return an access token that allows access to the server for some time period
-	public int login(int studentid, String password) throws 
+	public long login(int studentid, String password) throws 
 		UnauthorizedAccess, RemoteException;
 
 	// Return a summary list of Assessments currently available for this studentid
-	public List<String> getAvailableSummary(int token, int studentid) throws
+	public List<String> getAvailableSummary(long token, int studentid) throws
 		UnauthorizedAccess, NoMatchingAssessment, RemoteException;
 
 	// Return an Assessment object associated with a particular course code
-	public Assessment getAssessment(int token, int studentid, String courseCode) throws
+	public Assessment getAssessment(long token, int studentid, String courseCode) throws
 		UnauthorizedAccess, NoMatchingAssessment, RemoteException;
 
 	// Submit a completed assessment
-	public void submitAssessment(int token, int studentid, Assessment completed) throws 
+	public void submitAssessment(long token, int studentid, Assessment completed) throws 
 		UnauthorizedAccess, NoMatchingAssessment, RemoteException;
 
 }
